Реструктуризация проект делится на несколко приложений, в которых каждое имеет возмможность подключать другое приложение для каждого приложения установлен свой основной корневой путь. все пути от корневого пути прописаны в routes.js, подключенные контроллеры должны находится в controllers.js
тесты и мидлварки перенесены
создана папка tools, в которой лежат различные функции которые могут быть использованы из разных частей кода
 некоторые пути переименованы
 при сборке теперь собираются только hbs файлы и статика, все остальное запускается из src
