+++ src/blocks/general/header/nav/nav.hbs
@@ -5,7 +5,7 @@
 {{/if_eq}}
 {{>general-button text='Рейтинги' href='/rating' style='link'}}
 {{#if user}}
-        {{>general-button text='Мои Квесты' href='/myquests' style='link'}}
+        {{>general-button text='Мои Квесты' href='/quests/my' style='link'}}
         {{>general-button text='Профиль' href='/profile' style='link'}}
 
 {{else}}
