+++ src/blocks/main-page/main-page.hbs
@@ -7,7 +7,7 @@
 <div class="quests-block">
     <section class="quest-query">
         {{#if user}}
-            {{>general-button text='Создать квест' href='/newquest' style='primary'}}
+            {{>general-button text='Создать квест' href='/quests/new' style='primary'}}
         {{else}}
             <div></div>
         {{/if}}
