+++ src/blocks/my-quests-page/my-quests-page.hbs
@@ -1,4 +1,4 @@
-{{>general-button text='Создать квест' href='/newquest' style='primary'}}
+{{>general-button text='Создать квест' href='/quests/new' style='primary'}}
 <div class="tabs">
     {{> general-tab
         name="tabs"
