+++ src/blocks/quest-list/item/item.hbs
@@ -1,5 +1,5 @@
 <!-- htmllint attr-bans="false" -->
-<a class="quest-link" href="/quest/{{quest.id}}"></a>
+<a class="quest-link" href="/quests/{{quest.id}}"></a>
 <div class="img-container">
     {{#if_neq quest.progress undefined}}
         <div class="progressbar-wrapper">
